requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
aiohttp
pipecat-ai[daily,silero,openai]>=0.0.83
pydub
PyYAML>=6.0
pytest
requests
ruff
mypy
types-PyYAML
types-requests